refactor(files): Change folder names and cleanup code
