+++ backend/models/statsClass.js
@@ -47,4 +47,4 @@ class Stats {
     }
 }
 
-export default Stats
+export default Stats;
