+++ backend/src/llm_handler/llmClass.js
@@ -1,11 +1,9 @@
 import dotenv from 'dotenv';
 import OpenAI from "openai";
-import { llmTool, llmToolProperty } from '../llm_tools/toolClass.js';
 
 dotenv.config();
 
 const OPENAI_API_KEY = process.env.OPENAI_API_KEY;
-const OPENAI_MODEL = process.env.OPENAI_MODEL || 'gpt-4.1';
 
 if (!OPENAI_API_KEY) {
   console.error('Error: Missing OPENAI_API_KEY in environment');
