+++ frontend/src/components/ChatBackground.jsx
