+++ frontend/src/styles/index.css
@@ -52,7 +52,7 @@ body {
   display: flex;
   align-items: center;
   justify-content: center;
-  background: linear-gradient(rgba(0, 0, 0, 0.4), rgba(0, 0, 0, 0.4)), url('./assets/welcome-banner2.jpeg');
+  background: linear-gradient(rgba(0, 0, 0, 0.4), rgba(0, 0, 0, 0.4)), url('../assets/welcome-banner2.jpeg');
   background-size: cover;
   background-position: center;
   background-repeat: no-repeat;
